requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Data science libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Data processing
scipy==1.11.2
joblib==1.3.2

# Database and ORM
sqlalchemy==2.0.21
psycopg2-binary==2.9.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
httpx==0.25.0

# Code quality
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Utilities
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.2.1
